+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -190,6 +190,14 @@ export function withExtendedResolver(
     },
   };
 
+  // The vendored canary modules live inside /static/canary-full/node_modules
+  // Adding the `index.js` allows us to add this path as `originModulePath` to
+  // resolve the nested `node_modules` folder properly.
+  const canaryModulesPath = path.join(
+    require.resolve('@expo/cli/package.json'),
+    '../static/canary-full/index.js'
+  );
+
   let _universalAliases: [RegExp, string][] | null;
 
   function getUniversalAliases() {
@@ -715,9 +723,10 @@ export function withExtendedResolver(
         // Change the node modules path for react and react-dom to use the vendor in Expo CLI.
         /^(react|react\/.*|react-dom|react-dom\/.*)$/.test(moduleName)
       ) {
-        context.nodeModulesPaths = [
-          path.join(require.resolve('@expo/cli/package.json'), '../static/canary-full'),
-        ];
+        // Modifying the origin module path changes the starting Node module resolution path to this folder
+        context.originModulePath = canaryModulesPath;
+        // Hierarchical lookup has to be enabled for this to work
+        context.disableHierarchicalLookup = false;
       }
 
       if (isServerEnvironment(context.customResolverOptions?.environment)) {